[変更]Session APIに対応
