+++ Niconicome/Models/Domain/Niconico/Watch/WatchSession.cs
@@ -1,16 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Niconicome.Extensions.System;
-using Niconicome.Models.Domain.Niconico.Net.Json;
+using Niconicome.Models.Domain.Local.Addons.API.Hooks;
 using Niconicome.Models.Domain.Niconico.Video.Infomations;
 using Niconicome.Models.Helper.Result.Generic;
-using Niconicome.Models.Playlist;
 using Dmc = Niconicome.Models.Domain.Niconico.Dmc;
-using DmcRequest = Niconicome.Models.Domain.Niconico.Net.Json.WatchPage.DMC.Request;
-using DmcResponse = Niconicome.Models.Domain.Niconico.Net.Json.WatchPage.DMC.Response;
 using Utils = Niconicome.Models.Domain.Utils;
 using Video = Niconicome.Models.Domain.Niconico.Video.Infomations;
 
@@ -26,20 +22,6 @@ namespace Niconicome.Models.Domain.Niconico.Watch
         public IDomainVideoInfo? Video { get; }
         WatchSessionState State { get; }
         Task<Dmc::IStreamsCollection> GetAvailableStreamsAsync();
-    }
-
-    public interface IWatchSessionInfo
-    {
-        string DmcResponseJsonData { get; }
-        string ContentUrl { get; }
-        string SessionId { get; }
-    }
-
-    public interface IDmcDataHandler
-    {
-        DmcRequest::DmcPostData GetPostData(IDomainVideoInfo videoInfo);
-        Task<IWatchSessionInfo> GetSessionInfoAsync(DmcRequest::DmcPostData dmcPostData);
-        Task<IWatchSessionInfo> GetSessionInfoAsync(IDomainVideoInfo videoinfo);
     }
 
     public enum WatchSessionState
@@ -59,13 +41,14 @@ namespace Niconicome.Models.Domain.Niconico.Watch
     /// </summary>
     public class WatchSession : IWatchSession
     {
-        public WatchSession(IWatchInfohandler watchInfo, INicoHttp http, Utils::ILogger logger, IDmcDataHandler dmchandler, IWatchPlaylisthandler playlisthandler)
+        public WatchSession(IWatchInfohandler watchInfo, INicoHttp http, Utils::ILogger logger, IDmcDataHandler dmchandler, IWatchPlaylisthandler playlisthandler, IHooksManager hooksManager)
         {
             this.watchInfo = watchInfo;
             this.http = http;
             this.logger = logger;
             this.dmchandler = dmchandler;
             this.playlisthandler = playlisthandler;
+            this.hooksManager = hooksManager;
         }
 
         ~WatchSession()
@@ -85,9 +68,13 @@ namespace Niconicome.Models.Domain.Niconico.Watch
 
         private readonly INicoHttp http;
 
+        private readonly IHooksManager hooksManager;
+
         private IWatchSessionInfo? watchSessionInfo;
 
         #endregion
+
+        #region Props
 
         /// <summary>
         /// 動画情報
@@ -120,6 +107,10 @@ namespace Niconicome.Models.Domain.Niconico.Watch
         /// 状態
         /// </summary>
         public WatchSessionState State { get; private set; } = WatchSessionState.NotInitialized;
+
+        #endregion
+
+        #region Method
 
         /// <summary>
         /// 動画情報を取得する
@@ -196,21 +187,123 @@ namespace Niconicome.Models.Domain.Niconico.Watch
                 return;
             }
 
-            try
+            IAttemptResult<IWatchSessionInfo> result;
+
+            if (this.hooksManager.IsRegistered(HookType.SessionEnsuring))
             {
-                this.watchSessionInfo = await this.dmchandler.GetSessionInfoAsync(this.Video);
+                result = await this.EnsureSessionWithAddonAsync(this.Video);
             }
-            catch (Exception e)
+            else
+            {
+                result = await this.EnsureSessionDefaultAsync(this.Video);
+            }
+
+            if (!result.IsSucceeded || result.Data is null)
             {
                 this.State = WatchSessionState.SessionEnsuringFailure;
-                this.logger.Error("セッションの確立に失敗しました。", e);
                 return;
+            }
+            else
+            {
+                this.watchSessionInfo = result.Data;
             }
 
             this.IsSessionEnsured = true;
             this.State = WatchSessionState.OK;
             this.StartHeartBeat();
             this.logger.Log($"{nicoId}の視聴セッションを確立しました。");
+        }
+
+        /// <summary>
+        /// 取得可能なストリームの一覧を返す
+        /// </summary>
+        /// <returns></returns>
+        public async Task<Dmc::IStreamsCollection> GetAvailableStreamsAsync()
+        {
+            this.CheckSessionExpired();
+
+            if (this.watchSessionInfo is null) throw new InvalidCastException("セッション情報がnullのため、ストリームの一覧を取得できません。");
+            var availableStreams = new List<Dmc::IStreamInfo>();
+
+            await this.GetAvailableStreamsAsync(this.PlaylistUrl, availableStreams);
+
+            var streams = new Dmc::StreamsCollection();
+            streams.AddRange(availableStreams);
+
+            return streams;
+
+        }
+
+        /// <summary>
+        /// セッションを破棄する
+        /// </summary>
+        public void Dispose()
+        {
+            this.IsSessionEnsured = false;
+            this.IsSessionExipired = true;
+            GC.SuppressFinalize(this);
+        }
+
+        #endregion
+
+        #region private
+
+        /// <summary>
+        /// セッションを確立する
+        /// </summary>
+        /// <param name="video"></param>
+        /// <returns></returns>
+        private async Task<IAttemptResult<IWatchSessionInfo>> EnsureSessionDefaultAsync(IDomainVideoInfo video)
+        {
+            IWatchSessionInfo sessionInfo;
+
+            try
+            {
+                sessionInfo = await this.dmchandler.GetSessionInfoAsync(video);
+            }
+            catch (Exception e)
+            {
+                this.logger.Error("セッションの確立に失敗しました。", e);
+                return new AttemptResult<IWatchSessionInfo>() { Message = "セッションの確立に失敗しました。", Exception = e };
+            }
+
+            return new AttemptResult<IWatchSessionInfo>() { IsSucceeded = true, Data = sessionInfo };
+
+        }
+
+        /// <summary>
+        /// アドオンでセッションを確立する
+        /// </summary>
+        /// <param name="video"></param>
+        /// <returns></returns>
+        private async Task<IAttemptResult<IWatchSessionInfo>> EnsureSessionWithAddonAsync(IDomainVideoInfo video)
+        {
+            IAttemptResult<dynamic> result = await this.hooksManager.EnsureSessionAsync(video.RawDmcInfo);
+            if (!result.IsSucceeded || result.Data is null)
+            {
+                return new AttemptResult<IWatchSessionInfo>() { Message = $"セッションの確立に失敗しました。(詳細:{result.Message})", Exception = result.Exception };
+            }
+
+            try
+            {
+                if (result.Data.DmcResponseJsonData is not string jsonData) return new AttemptResult<IWatchSessionInfo>() { Message = $"HeratBeatコンテントが正しくありません。" };
+                if (result.Data.ContentUrl is not string contentUrl) return new AttemptResult<IWatchSessionInfo>() { Message = $"コンテンツURLが正しくありません。" };
+                if (result.Data.SessionId is not string sessionID) return new AttemptResult<IWatchSessionInfo>() { Message = $"セッションIDが正しくありません。" };
+
+                var info = new WatchSessionInfo()
+                {
+                    DmcResponseJsonData = result.Data.DmcResponseJsonData,
+                    ContentUrl = result.Data.ContentUrl,
+                    SessionId = result.Data.SessionId,
+                };
+
+                return new AttemptResult<IWatchSessionInfo>() { IsSucceeded = true, Data = info };
+            }
+            catch (Exception e)
+            {
+                this.logger.Error("アドオンから返却されたオブジェクトの解析に失敗しました。", e);
+                return new AttemptResult<IWatchSessionInfo>() { Message = $"アドオンから返却されたオブジェクトの解析に失敗しました。", Exception = e };
+            }
         }
 
         /// <summary>
@@ -250,36 +343,6 @@ namespace Niconicome.Models.Domain.Niconico.Watch
                 }
                 streams.Add(stream);
             }
-        }
-
-        /// <summary>
-        /// 取得可能なストリームの一覧を返す
-        /// </summary>
-        /// <returns></returns>
-        public async Task<Dmc::IStreamsCollection> GetAvailableStreamsAsync()
-        {
-            this.CheckSessionExpired();
-
-            if (this.watchSessionInfo is null) throw new InvalidCastException("セッション情報がnullのため、ストリームの一覧を取得できません。");
-            var availableStreams = new List<Dmc::IStreamInfo>();
-
-            await this.GetAvailableStreamsAsync(this.PlaylistUrl, availableStreams);
-
-            var streams = new Dmc::StreamsCollection();
-            streams.AddRange(availableStreams);
-
-            return streams;
-
-        }
-
-        /// <summary>
-        /// セッションを破棄する
-        /// </summary>
-        public void Dispose()
-        {
-            this.IsSessionEnsured = false;
-            this.IsSessionExipired = true;
-            GC.SuppressFinalize(this);
         }
 
         /// <summary>
@@ -332,175 +395,7 @@ namespace Niconicome.Models.Domain.Niconico.Watch
                 throw new InvalidOperationException("セッションが失効しているため操作を完了できません。オブジェクトを再生成して下さい。");
             }
         }
-    }
 
-    /// <summary>
-    /// セッション情報
-    /// </summary>
-    public class WatchSessionInfo : IWatchSessionInfo
-    {
-        public string DmcResponseJsonData { get; set; } = string.Empty;
-        public string ContentUrl { get; set; } = string.Empty;
-        public string SessionId { get; set; } = string.Empty;
-
-    }
-
-    /// <summary>
-    /// DMCサーバーとの通信を管理する
-    /// </summary>
-    public class DmcDataHandler : IDmcDataHandler
-    {
-        public DmcDataHandler(INicoHttp http)
-        {
-            this.http = http;
-        }
-
-        /// <summary>
-        /// httpクライアント
-        /// </summary>
-        private readonly INicoHttp http;
-        public DmcRequest::DmcPostData GetPostData(IDomainVideoInfo videoInfo)
-        {
-            var data = DmcRequest::DmcPostData.GetInstance();
-            var sessionIfnfo = videoInfo.DmcInfo.SessionInfo;
-
-            //nullチェック
-            var nullList = new List<string>();
-            var type = sessionIfnfo.GetType();
-            var properties = type.GetProperties();
-            foreach (var prop in properties)
-            {
-                string name = prop.Name;
-                object? value = prop.GetValue(sessionIfnfo);
-                if (value is null)
-                {
-                    nullList.Add(name);
-                }
-            }
-
-            if (nullList.Count > 0)
-            {
-                throw new InvalidOperationException($"DMCサーバーへのPOSTに必要なデータがnullです。(一覧: {string.Join(',', nullList)})");
-            }
-
-            data.Session.Recipe_id = sessionIfnfo.RecipeId;
-            data.Session.Content_id = sessionIfnfo.ContentId;
-            data.Session.Content_type = "movie";
-            data.Session.Content_src_id_sets.Add(this.GetContentSrcIdSets(sessionIfnfo));
-            data.Session.Timing_constraint = "unlimited";
-            data.Session.Keep_method.Heartbeat.Lifetime = sessionIfnfo.HeartbeatLifetime;
-            data.Session.Protocol.Name = "http";
-            data.Session.Protocol.Parameters.Http_parameters.Parameters.Hls_parameters.Use_ssl = "yes";
-            data.Session.Protocol.Parameters.Http_parameters.Parameters.Hls_parameters.Use_well_known_port = "yes";
-            data.Session.Protocol.Parameters.Http_parameters.Parameters.Hls_parameters.Transfer_preset = sessionIfnfo.TransferPriset;
-            data.Session.Protocol.Parameters.Http_parameters.Parameters.Hls_parameters.Segment_duration = 6000;
-            data.Session.Session_operation_auth.Session_operation_auth_by_signature.Signature = sessionIfnfo.Signature;
-            data.Session.Session_operation_auth.Session_operation_auth_by_signature.Token = sessionIfnfo.Token;
-            data.Session.Content_auth.Auth_type = sessionIfnfo.AuthType;
-            data.Session.Content_auth.Service_id = "nicovideo";
-            data.Session.Content_auth.Service_user_id = sessionIfnfo.ServiceUserId;
-            data.Session.Content_auth.Content_key_timeout = sessionIfnfo.ContentKeyTimeout;
-            data.Session.Client_info.Player_id = sessionIfnfo.PlayerId;
-            data.Session.Priority = sessionIfnfo.Priority;
-
-
-            return data;
-        }
-
-        /// <summary>
-        /// セッション情報を取得する
-        /// </summary>
-        /// <param name="dmcPostData"></param>
-        /// <returns></returns>
-        public async Task<IWatchSessionInfo> GetSessionInfoAsync(DmcRequest::DmcPostData dmcPostData)
-        {
-
-            string json = JsonParser.Serialize(dmcPostData);
-
-            var res = await this.http.PostAsync(new Uri("https://api.dmc.nico/api/sessions?_format=json"), new StringContent(json));
-            if (!res.IsSuccessStatusCode)
-            {
-                throw new HttpRequestException($"DMCサーバーへのPOSTに失敗しました。(status: {(int)res.StatusCode}, reason_phrase: {res.ReasonPhrase})");
-            }
-
-            string responseString = await res.Content.ReadAsStringAsync();
-
-            var deserialized = JsonParser.DeSerialize<DmcResponse::DmcResponseData>(responseString);
-
-            if (deserialized.Data.Session.ContentUri.IsNullOrEmpty())
-            {
-                throw new HttpRequestException("DMCサーバーからの情報取得に失敗しました。(Data.Session.ContentUriがnullまたは空白です。)");
-            }
-            else if (deserialized.Data.Session.Id.IsNullOrEmpty())
-            {
-                throw new HttpRequestException("DMCサーバーからの情報取得に失敗しました。(Data.Idがnullまたは空白です。)");
-            }
-
-            return new WatchSessionInfo()
-            {
-                DmcResponseJsonData = JsonParser.Serialize(deserialized.Data),
-                ContentUrl = deserialized.Data.Session.ContentUri,
-                SessionId = deserialized.Data.Session.Id,
-            };
-        }
-
-        /// <summary>
-        /// セッション情報を取得する
-        /// </summary>
-        /// <param name="videoinfo"></param>
-        /// <returns></returns>
-        public Task<IWatchSessionInfo> GetSessionInfoAsync(IDomainVideoInfo videoinfo)
-        {
-            var postData = this.GetPostData(videoinfo);
-            return this.GetSessionInfoAsync(postData);
-        }
-
-        #region private
-
-
-        /// <summary>
-        /// Content_Src_Id_Setsを構成する
-        /// </summary>
-        /// <param name="sessionInfo"></param>
-        /// <returns></returns>
-        private DmcRequest::Content_Src_Id_Sets GetContentSrcIdSets(ISessionInfo sessionInfo)
-        {
-
-            sessionInfo.Videos.Sort((a, b) =>
-            {
-                if (a.EndsWith("_low")) return -1;
-                if (b.EndsWith("_low")) return 1;
-                return string.Compare(a, b);
-            });
-            var videoSrc = sessionInfo.Videos.Select((value, index) => new { value, index }).ToList();
-            string audio = sessionInfo.Audios[0];
-            var sets = new DmcRequest::Content_Src_Id_Sets();
-
-            foreach (var video in videoSrc)
-            {
-                var idsData = new DmcRequest::Content_Src_Ids();
-                idsData.Src_id_to_mux.Audio_src_ids.Add(audio);
-                int videosCount = videoSrc.Count - video.index;
-
-                foreach (var i in Enumerable.Range(0, videosCount))
-                {
-                    idsData.Src_id_to_mux.Video_src_ids.Add(videoSrc[i].value);
-                }
-
-                //idsData.Src_id_to_mux.Video_src_ids = idsData.Src_id_to_mux.Video_src_ids.OrderByDescending(s => s).ToList();
-                idsData.Src_id_to_mux.Video_src_ids.Sort((a, b) =>
-                {
-                    if (a.EndsWith("_low")) return 1;
-                    if (b.EndsWith("_low")) return -1;
-                    return string.Compare(b, a);
-                });
-
-                sets.Content_src_ids.Add(idsData);
-
-            }
-
-            return sets;
-        }
         #endregion
     }
